Allow task list checkboxes to be deeply nested
We previously only detected the leading checkbox in a task list item if it was the first element or if the first element was a `<p>` and the checkbox was the first element in it. This change allows the checkbox to be more deeply nested inside other phrasing elements, like `<em>` or `<strong>`, so long as it is the first thing in them. For example:

    <li>
      <p>
        <strong>
          <ins>
            <input type="checkbox">
            Some other text
          </ins>
        </strong>
      </p>
    </li>